Debugging of diffusion code and initialization updates
Fixed (at least mostly) so code now works on meshes containing elements of unequal volumes. There may still be a bug in the vasculature growth/movement as well as in calculating the target value for the source term in the secretion term in the diffusion code. The weighting originally was incorrect, but the correction currently in place needs looked at.