Setup PC for redoing angiogenesis code
Getting seg fault or bad alloc consistently. Seems to be isolated to custom codes. Debugging was attempted but did not resolve problem. Will rebuild the custom codes additively in the hopes of either finding the error and getting around it or by recoding not having an error. Will use previous code (backups) as a starting point for new codes.